+++ src/Data/Morpheus/Execution/Document/Declare.hs
@@ -1,5 +1,6 @@
-{-# LANGUAGE DataKinds      #-}
-{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE DataKinds       #-}
+{-# LANGUAGE NamedFieldPuns  #-}
+{-# LANGUAGE TemplateHaskell #-}
 
 module Data.Morpheus.Execution.Document.Declare
   ( declareTypes
@@ -11,14 +12,18 @@ import           Language.Haskell.TH
 --
 -- MORPHEUS
 import           Data.Morpheus.Execution.Document.GQLType (deriveGQLType)
-import           Data.Morpheus.Execution.Internal.Declare (declareType)
-import           Data.Morpheus.Types.Internal.DataD       (GQLTypeD)
+import           Data.Morpheus.Execution.Internal.Declare (declareResolverType, declareType)
+import           Data.Morpheus.Types.Internal.DataD       (GQLTypeD (..), isInputKind)
 
 declareTypes :: [GQLTypeD] -> Q [Dec]
 declareTypes = fmap concat . traverse declareGQLType
 
 declareGQLType :: GQLTypeD -> Q [Dec]
-declareGQLType gqlType@(typeD, _, argTypes) = do
-  let types = map (declareType []) (typeD : argTypes)
+declareGQLType gqlType@GQLTypeD {typeD, typeKindD, typeArgD} = do
+  let types = declareResolverType typeKindD derivingClasses typeD : map (declareType []) typeArgD
   typeClasses <- deriveGQLType gqlType
   pure $ types <> typeClasses
+  where
+    derivingClasses
+      | isInputKind typeKindD = [''Show]
+      | otherwise = []