+++ src/Data/Morpheus/Execution/Document/GQLType.hs
@@ -15,13 +15,30 @@ import           Data.Morpheus.Kind                 (ENUM, INPUT_OBJECT, INPUT_U
 -- MORPHEUS
 import           Data.Morpheus.Types.GQLType        (GQLType (..))
 import           Data.Morpheus.Types.Internal.Data  (DataTypeKind (..))
-import           Data.Morpheus.Types.Internal.DataD (GQLTypeD, TypeD (..))
+import           Data.Morpheus.Types.Internal.DataD (GQLTypeD (..), KindD (..), TypeD (..), unKindD)
+import           Data.Typeable                      (Typeable)
 
 deriveGQLType :: GQLTypeD -> Q [Dec]
-deriveGQLType (TypeD {tName}, gqlKind, _) =
-  pure <$> instanceD (cxt []) (appT (conT ''GQLType) (conT $ mkName tName)) methods
+deriveGQLType GQLTypeD {typeD = TypeD {tName}, typeKindD} =
+  pure <$> instanceD (cxt constrains) (appT (conT ''GQLType) genHeadSig) [methods]
   where
-    methods = [pure $ TySynInstD ''KIND (TySynEqn [ConT $ mkName tName] (ConT $ toKIND gqlKind))]
+    gqlKind = unKindD typeKindD
+    withVar = gqlKind == KindObject || gqlKind == KindUnion
+    isSubscription = typeKindD == SubscriptionD
+    genHeadSig
+      | isSubscription = appT (appT (conT $ mkName tName) (varT $ mkName "subscriptionM")) (varT $ mkName "m")
+      | withVar = appT (conT $ mkName tName) (varT $ mkName "m")
+      | otherwise = conT $ mkName tName
+    ----------
+    constrains
+      | isSubscription = map consTypeable ["subscriptionM", "m"]
+      | withVar = [consTypeable "m"]
+      | otherwise = []
+    consTypeable = appT (conT ''Typeable) . (varT . mkName)
+    ----
+    methods = do
+      typeN <- genHeadSig
+      pure $ TySynInstD ''KIND (TySynEqn [typeN] (ConT $ toKIND gqlKind))
     toKIND KindScalar      = ''SCALAR
     toKIND KindEnum        = ''ENUM
     toKIND KindObject      = ''OBJECT