+++ src/Data/Morpheus/Types.hs
@@ -9,6 +9,7 @@ module Data.Morpheus.Types
   , IOSubRes
   , Resolver
   , SubRootRes
+  , SubResolver(..)
   , Event(..)
   -- Type Classes
   , GQLType(KIND, description)
@@ -19,6 +20,7 @@ module Data.Morpheus.Types
   , ID(..)
   , ScalarValue(..)
   , GQLRootResolver(..)
+  , constRes
   ) where
 
 import           Data.Morpheus.Types.GQLScalar      (GQLScalar (parseValue, serialize))
@@ -27,10 +29,14 @@ import           Data.Morpheus.Types.ID             (ID (..))
 import           Data.Morpheus.Types.Internal.Value (ScalarValue (..))
 import           Data.Morpheus.Types.IO             (GQLRequest (..), GQLResponse (..))
 import           Data.Morpheus.Types.Resolver       (Event (..), GQLRootResolver (..), MutResolver, Resolver,
-                                                     SubResolver, SubRootRes, mutResolver, resolver, toMutResolver)
+                                                     SubResolver (..), SubRootRes, mutResolver, resolver, toMutResolver)
+
+-- resolves constant value on any argument
+constRes :: Monad m => b -> a -> m b
+constRes = const . return
 
 type IORes = Resolver IO
 
 type IOMutRes e c = MutResolver IO e c
 
-type IOSubRes e c a = SubResolver IO e c a
+type IOSubRes e c = SubResolver IO e c