+++ test/Feature/Holistic/API.hs
@@ -1,9 +1,11 @@
-{-# LANGUAGE DeriveGeneric     #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE QuasiQuotes       #-}
-{-# LANGUAGE TemplateHaskell   #-}
-{-# LANGUAGE TypeFamilies      #-}
-{-# LANGUAGE TypeOperators     #-}
+{-# LANGUAGE DeriveGeneric      #-}
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE NamedFieldPuns     #-}
+{-# LANGUAGE OverloadedStrings  #-}
+{-# LANGUAGE QuasiQuotes        #-}
+{-# LANGUAGE TemplateHaskell    #-}
+{-# LANGUAGE TypeFamilies       #-}
+{-# LANGUAGE TypeOperators      #-}
 
 module Feature.Holistic.API
   ( api
@@ -12,17 +14,15 @@ module Feature.Holistic.API
 import           Data.Morpheus          (interpreter)
 import           Data.Morpheus.Document (importGQLDocument)
 import           Data.Morpheus.Kind     (SCALAR)
-import           Data.Morpheus.Types    (Event (..), GQLRequest, GQLResponse, GQLRootResolver (..), GQLScalar (..),
-                                         GQLType (..), ID (..), IORes, IOSubRes, ScalarValue (..))
+import           Data.Morpheus.Types    (GQLRequest, GQLResponse, GQLRootResolver (..), GQLScalar (..), GQLType (..),
+                                         ID (..), IOMutRes, IORes, IOSubRes, ScalarValue (..), SubResolver (..))
 import           Data.Text              (Text)
 import           GHC.Generics           (Generic)
-
-importGQLDocument "test/Feature/Holistic/API.gql"
 
 data TestScalar =
   TestScalar Int
              Int
-  deriving (Generic)
+  deriving (Show, Generic)
 
 instance GQLType TestScalar where
   type KIND TestScalar = SCALAR
@@ -31,42 +31,37 @@ instance GQLScalar TestScalar where
   parseValue _ = pure (TestScalar 1 0)
   serialize (TestScalar x y) = Int (x * 100 + y)
 
-newtype Mutation = Mutation
-  { createUser :: AddressArgs -> IORes User
-  } deriving (Generic)
-
 data EVENT =
   EVENT
   deriving (Show, Eq)
 
-newtype Subscription = Subscription
-  { newUser :: AddressArgs -> IOSubRes EVENT () User
-  } deriving (Generic)
+importGQLDocument "test/Feature/Holistic/API.gql"
 
 resolveValue :: Monad m => b -> a -> m b
 resolveValue = const . return
 
-resolveUser :: a -> IORes User
-resolveUser _ =
-  return $
-  User
-    { name = resolveValue "testName"
-    , email = resolveValue ""
-    , address = resolveAddress
-    , office = resolveAddress
-    , friend = resolveValue Nothing
-    }
-  where
-    resolveAddress _ =
-      return Address {city = resolveValue "", houseNumber = resolveValue 0, street = resolveValue Nothing}
-
-rootResolver :: GQLRootResolver IO EVENT () Query Mutation Subscription
+rootResolver ::
+     GQLRootResolver IO EVENT () (Query IORes) (Mutation (IOMutRes EVENT ())) (Subscription (IOSubRes EVENT ()) IORes)
 rootResolver =
   GQLRootResolver
-    { queryResolver = return Query {user = resolveUser, testUnion = const $ return Nothing}
-    , mutationResolver = return Mutation {createUser = resolveUser}
-    , subscriptionResolver = return Subscription {newUser = const $ Event [EVENT] resolveUser}
+    { queryResolver = return Query {user, testUnion = const $ return Nothing}
+    , mutationResolver = return Mutation {createUser = user}
+    , subscriptionResolver =
+        return Subscription {newUser = const SubResolver {subChannels = [EVENT], subResolver = user}}
     }
+  where
+    user _ =
+      return $
+      User
+        { name = resolveValue "testName"
+        , email = resolveValue ""
+        , address = resolveAddress
+        , office = resolveAddress
+        , friend = resolveValue Nothing
+        }
+      where
+        resolveAddress _ =
+          return Address {city = resolveValue "", houseNumber = resolveValue 0, street = resolveValue Nothing}
 
 api :: GQLRequest -> IO GQLResponse
 api = interpreter rootResolver